requirements.txt
numpy==1.23.0
pandas==1.5.0
matplotlib==3.7.1
seaborn==0.11.2
scikit-learn==1.1.2
streamlit==1.12.2
python-docx==0.8.11
PyPDF2==2.11.2